pyproject.toml
[tool.poetry]
name = "test-iqeq"
version = "0.1.0"
description = ""
authors = ["artemiikulikov <aapqwhat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
django = "~3.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
